+++ server/src/main/java/org/allaymc/server/entity/component/EntityBaseComponentImpl.java
@@ -324,7 +324,7 @@ public class EntityBaseComponentImpl implements EntityBaseComponent {
     }
 
     @Override
-    public void despawn() {
+    public void remove() {
         getDimension().getEntityService().removeEntity(thisEntity);
     }