+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityXpBottlePhysicsComponentImpl.java
@@ -27,7 +27,7 @@ public class EntityXpBottlePhysicsComponentImpl extends EntityProjectilePhysicsC
             return;
         }
 
-        thisEntity.despawn();
+        thisEntity.remove();
         spawnExp();
     }